requirements.txt
vastai-sdk>=0.4.2
